CM_API_Training-master/requirements.txt
requests==2.18.4
PyJWT==1.6.1
